requirements.txt
streamlit==1.30.0
google-generativeai==0.4.0
python-dotenv==1.0.0
